isolate getGroupedMixinOptions
