+++ src/editor/components/widgets/Mixins.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import Select, { components } from 'react-select';
 import Events from '../../lib/Events';
 import { DropdownArrowIcon } from '../../icons';
+import { getGroupedMixinOptions } from '../../lib/mixinUtils';
 
 export default class Mixin extends React.Component {
   static propTypes = {
@@ -44,31 +45,6 @@ export default class Mixin extends React.Component {
       .filter((v) => !!v)
       .map((v) => ({ label: v, value: v }));
   }
-
-  getGroupedMixinOptions = () => {
-    const mixinElements = document.querySelectorAll('a-mixin');
-    const groupedArray = [];
-    let categoryName, mixinId;
-
-    const groupedObject = {};
-    for (let mixinEl of Array.from(mixinElements)) {
-      categoryName = mixinEl.getAttribute('category');
-      if (!categoryName) continue;
-      mixinId = mixinEl.id;
-      if (!groupedObject[categoryName]) {
-        groupedObject[categoryName] = [];
-      }
-      groupedObject[categoryName].push({ label: mixinId, value: mixinId });
-    }
-
-    for (let categoryName of Object.keys(groupedObject)) {
-      groupedArray.push({
-        label: categoryName,
-        options: groupedObject[categoryName]
-      });
-    }
-    return groupedArray;
-  };
 
   updateMixins = (value) => {
     const entity = this.props.entity;
@@ -136,7 +112,7 @@ export default class Mixin extends React.Component {
               <Select
                 id="mixinSelect"
                 classNamePrefix="select"
-                options={this.getGroupedMixinOptions()}
+                options={getGroupedMixinOptions(false)}
                 components={{
                   GroupHeading: CustomGroupHeading,
                   DropdownIndicator: DropdownArrowIcon,
@@ -158,7 +134,7 @@ export default class Mixin extends React.Component {
               <Select
                 id="mixinSelect"
                 classNamePrefix="select-single"
-                options={this.getGroupedMixinOptions()}
+                options={getGroupedMixinOptions(false)}
                 components={{
                   GroupHeading: CustomGroupHeading,
                   DropdownIndicator: DropdownArrowIcon,
